requirements.txt
Django==4.2.11
celery==5.3.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pandas
pytz
djangorestframework-simplejwt
redis==5.0.3
requests==2.31.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-celery-beat
paho-mqtt
xlrd==2.0.1